Use CentOS 7 as base
Previous to this commit, the image was based on a ruby:2.4 which is
based on Alpine. This caused us to have to install PDK as a gem, which
isn't ideal. We also had to compile an older version of curl which
introduced CVEs. The base Ruby image also contained CVEs.

This commit moves everything over to CentOS 7 and installs PDK as a RPM,
enabling PDK's full capabilities.