Compile ra10ke gem from source
Previous to this commit, the Puppetfile syntax checks would always pass,
even if the Puppetfile contained a syntax error. This was because the
ra10ke gem's r10k:syntax rake task always exited zero.

The issue has been fixed in master and this commit compiles from master
to get the fix until it is released to the gem