Provide a way to exclude syntax/lint paths
Previous to this commit, the list of paths to exclude from linting and
parser validation was static. For users that have Bolt plans, for
example, that need to be excluded from the syntax and linting checks,
there was no way to do so.

This commit provides the EXCLUDE_PATHS environment variable, deliniated
by ':' characters, to provide a list of paths to exclude from linting
and syntax checks.

For example: `export EXCLUDE_PATHS=modules/**/plans/*:site/modules/**/bad_code/*`