Adapt GH Action to also push to Docker Hub
Part of this commit includes switching from two Dockerfiles to a single
one that does a multi-stage build. All the same images are built but now
the gosu image is guaranteed to be built off of the main image that was
also just built. Part of the need for the change to what gosu uses as a
`FROM` is an additional change that removes the user from the images
being built so that when they are build by the GH Action the username
can be configurable. This allows for anyone who has forked the repo to
also build Docker images and push them to their own account on Docker
Hub.

You may also notice that the end of the Dockerfile effectively takes
`base` and renames it to `main`. This is done so that two things can be
simultaneously true:

1. the gosu image is built off of the base image
2. if `--target` is not specified the result is the "main" image...
    which happens to be the exact same as `base`.

The following additional changes are also included:

- A minor edit was made to the Dockerfile to cleanup the bundler
  config set at build time so as to be sure not to bust any use of
  bundler by the end user.
- A couple of minor edits were made to the Gemfile to improve
  consistency in light of the fact that Dependabot is now keeping this
  file up to date.
- The README was updated to document what tags are pushed to Docker Hub
- Microbadger badges were added to show latest version and commit on
  Docker Hub
- A couple of pieces of metadata were added into the Dockerfile to be
  included in the images as labels. These labels power the Microbadger
  badges.