Update apt cache before upgrade
Previous to this commit, the apt cache was not updated prior to doing an
apt package upgrade. This caused important updates to be missed. This
commit introduces an initial apt cache update prior to doing an apt
upgrade. Another apt cache update is done after the puppet agent
repository package is installed.