(CDPE-4329) Github action to promote 4.x image
With this commit, we add a Github action to take a dev image (and
rootless dev image) and promote them to 4.x and 4.x-rootless,
respectively. This action is to be used by automation in the Pipelines
Infra repo to promote test images to the 4.x tag before each release
in order to release puppet-dev-tools monthly. See
https://tickets.puppetlabs.com/browse/CDPE-4314 for more details.