(CDPE-4471) Remove remaining unneeded files from test module
WIth this commit, we remove all unneeded files from the pdk 1.18.1 test
module committed to this repository. The only suspect remaining file is
the rubocop config, as without it, "pdk validate" will fail with rubocop
errors. Its likely possible we could commit a simplified version of the
rubocop file to this repository instead of the entire file from the PDK,
but doing so would require more trial/error than its currently worth.