(CDPE-4856) Remove more puppet gem keys in Dockerfile
This also adds a splat to our removal of puppet-7.x.x test fixtures in
the Dockerfile since we don't want to try and keep up with every y release of the
puppet gem. This removal comes from another customer ask.

Also, we are removing a bunch of files (at the behest of aforementioned customer) in a
separate layer at the end of our Dockerfile. TIL it's not that huge of a lift to
go back a few layers and grab those files from the docker image.
So. This moves the removal commands into the same layers that create the
files so that they are never present in the image, even on layer introspection.