(maint) Update critical security packages
We were encountering a couple CVEs with the 2.7.4-slim-buster image (CVE-2022-29155, CVE-2022-1292) around ldap and ssl. Instead of upgrading the base image, this just upgrades the specific offending packages. We don't have a lot of ways to be confident about a base image bump right now, so this is safer.