(maint) Add Trivy to 4.x publish
This adds a Trivy scan to the 4.x publish action. Since vulnerabilities
can be discovered at any time, we want to double check the dev image before we
publish to 4.x. As part of adding the Trivy step, this also splits out
the docker pull part of the action, which saves time on the Trivy scan that was timing out.