(maint) Skip scanning nokogiri's deps file
This deps file contains an RSA key that was used to for GPG signing of
one of the deps. Trivy is incorrectly flagging it as a leaked AWS key.
All future versions of nokogiri contain this key in their deps file, so
we are just skipping the file when doing the Trivy scan, since it is not
a real security issue.