(CDPE-6528) Input image tag name to Build-Test-Push
Add optional input to the Build-Test-Push job for supplying an image tag
to create in addition to the latest and timestamp-sha tags.

This enables workflows to call this and reliable know the name of the
resulting image, and is a result of that fact that GitHub does not have
a native way to run a workflow on a remote repository and get some sort
of string-like output from it.

Since there's no way to get data back from the called workflow, instead
let the calling workflow decide what name to use for the image tag.

Overall this enables this workflow to be called in automation by the
CD4PE release workflows so the developer doing the CD4PE release doesn't
have to manually trigger this first, find the image tag it created, then
supply that image tag to a different workflow.