Update Build-Test-Push job to publish `puppet8` images
Previously this job would publish "latest" and timestamped images, but
when producing images with Puppet 8 tooling, we don't want these images,
as they are easily confused with the versions produced from `main`.
Instead, this job should be used to release a new `puppet8` image, by
triggering it from the `puppet8` branch of the repo.