(CDPE-7069) Start caching trivy DBs in separate workflow
In order to avoid ratelimiting issues when running trivy scans,
all trivy runs need to run against a local copy of
the DB (rather than downloading it each time). This commit
adds a workflow to download the trivy DBs each morning.

Once this is in place and runs at least once, the trivy scans can be
added back in and set to use the cache rather than download.

This PR is essentially following the documented steps at
https://github.com/aquasecurity/trivy-action?tab=readme-ov-file#updating-caches-in-the-default-branch