(maint) Update to Ruby 3.2, PDK 3.4.0, and Debian bullseye
This commit updates PDK to current latest, 3.4.0. It also removes some
unnecessary `rm` lines from the Dockerfile: in PDK 3.2.0 the
dependencies that used to be used to analytics were dropped when the
analytics code was removed, and this included the `httpclient` gem,
which contained the files we used to delete because they triggered
security scanners. The gem is no longer included with the PDK package,
so the delete lines aren't needed.

Building this image with the latest nokogiri (1.18.8) ran into an issue
where they have split their gem platforms for the precompiled gem from
just `linux` to `linux-gnu` and `linux-musl`. However on Ruby 3.1 and
Debian Buster, bundler recognized the platform as just plain `linux`, so
it failed to grab the precompiled gems. We wanted to update to Ruby 3.2
to match what puppet-agent ships with anyway, and the docker images for
3.2 only support back through bullseye, not buster, so this commit
also updates to Debian Buster and Ruby 3.2. On this platform, bundler
correctly detects the gem platform and installs the correct pre-compiled
gems.

Updating to Ruby 3.2 required installing a newer version of rugged (a
dependency of octocatalog-diff), and this version depends on
`pkg-config`, as well as `libssl-dev` rather than just `libssl1.1` for
HTTPS support, so this commit also installs those.